add step over functionality, but the UI behaves nastily
For all the stepping operations, the currently selected stack frame is relevant.
However, that frame is not reliably communicated to the underlying Debugger when
a) the package filter changes
b) a frame is clicked twice (usually that unselects the frame, not so in Vivide lists).
Also, when step over is clicked and as a result,
multiple stack frames vanish from above,
the selection will still be at the nth frame from the top like before
and therefore the view will have jumped to a different frame
than the one in which the step was done.
Also, the method source code brick does not like to adhere to its script,
but instead always changes back to the Inspector, without explanation so far.